refactor extract formatBar functions (useFormatUtils)
- create customHook useFormatUtils.ts
- extract Format&FormatBar functions to useFormatUtils.ts
