--- frontend/src/utils/format.ts
@@ -1,6 +1,0 @@
-export enum FormatType {
-	BOLD = "bold",
-	ITALIC = "italic",
-	CODE = "code",
-	STRIKETHROUGH = "strikeThrough",
-}
